fix(voice): stop double-response bug; move voice toggle into ManyChat; drop /multi-agent
The live test surfaced two issues:

1. Bug — "Conversation already has an active response in progress". useVoiceSession
   ran handleToolCall on BOTH response.function_call_arguments.done AND
   response.output_item.done for the same call_id, sending the tool output + a
   response.create twice; the second collided with the in-flight response. Fix:
   de-dupe tool calls by call_id, track an active-response flag (set on
   response.created / cleared on response.done), and defer response.create until
   the active response finishes.

2. Wrong surface — voice was wired into Chat.tsx (legacy: /multi-agent stub +
   commented-out in ProjectContent), not the real chat. Move the 🎙 toggle into
   ManyChat, which renders in the global Zoe drawer (Cmd+J, every page) AND on
   /agent. Spoken turns render in the thread marked 🎙 (ChatMessage.marker),
   workspace-scoped, with end-on-silence, manual close, and mic-permission UX.
   Revert Chat.tsx to its pre-voice state.

Also delete the unused /multi-agent route (+ MultiChat stub, middleware/robots
refs) per request. /voice-debug kept as a dev/QA tool.
